+++ src/app/(pages)/experience/add/_components/ExperienceAddNotionPageSelectView.tsx
@@ -2,6 +2,12 @@ import { CheckedBoxIcon } from '@/components/common/icons/CheckedBoxIcon';
 import { ChevronLeftIcon } from '@/components/common/icons/ChevronLeftIcon';
 import { EmptyBoxIcon } from '@/components/common/icons/EmptyBoxIcon';
 import { NotionIcon } from '@/components/common/icons/NotionIcon';
+import {
+  formatLastEditedTime,
+  getNotionTagTone,
+  getNotionTypeLabel,
+  isUrlIcon,
+} from '@/app/(pages)/experience/add/_utils/notionPageDisplay';
 import { ModalClose, ModalDescription, ModalTitle } from '@/components/common/Modal';
 import { Tag } from '@/components/common/Tag';
 import { Button } from '@/components/ui/button';
@@ -168,34 +174,4 @@ function NotionPageIcon({ icon }: { icon?: string | null }) {
   }
 
   return <span className="text-xl leading-none">{icon}</span>;
-}
-
-function isUrlIcon(icon: string) {
-  return icon.startsWith('http://') || icon.startsWith('https://');
-}
-
-function getNotionTypeLabel(type?: string | null) {
-  return type === 'database' ? '데이터베이스' : '페이지';
-}
-
-function getNotionTagTone(type?: string | null): React.ComponentProps<typeof Tag>['tone'] {
-  return type === 'database' ? 'skill' : 'competency';
-}
-
-function formatLastEditedTime(lastEditedTime?: string | null) {
-  if (!lastEditedTime) {
-    return '최근 수정일 정보 없음';
-  }
-
-  const date = new Date(lastEditedTime);
-
-  if (Number.isNaN(date.getTime())) {
-    return lastEditedTime;
-  }
-
-  const year = date.getFullYear();
-  const month = String(date.getMonth() + 1).padStart(2, '0');
-  const day = String(date.getDate()).padStart(2, '0');
-
-  return `${year}.${month}.${day}`;
 }
