+++ src/app/(pages)/experience/add/_components/ExperienceAddUploadStep.tsx
@@ -13,6 +13,12 @@ import {
   clearExperienceAddPdfDraft,
   saveExperienceAddPdfDraft,
 } from '@/app/(pages)/experience/add/_utils/experienceAddPdfDraftStorage';
+import {
+  formatLastEditedTime,
+  getNotionTagTone,
+  getNotionTypeLabel,
+  isUrlIcon,
+} from '@/app/(pages)/experience/add/_utils/notionPageDisplay';
 import { EmptyState } from '@/components/common/EmptyState';
 import { Modal } from '@/components/common/Modal';
 import { Tag } from '@/components/common/Tag';
@@ -237,34 +243,4 @@ function NotionMaterialIcon({ icon }: { icon?: string | null }) {
   }
 
   return <span className="text-xl leading-none">{icon}</span>;
-}
-
-function isUrlIcon(icon: string) {
-  return icon.startsWith('http://') || icon.startsWith('https://');
-}
-
-function getNotionTypeLabel(type?: string | null) {
-  return type === 'database' ? '데이터베이스' : '페이지';
-}
-
-function getNotionTagTone(type?: string | null): React.ComponentProps<typeof Tag>['tone'] {
-  return type === 'database' ? 'skill' : 'competency';
-}
-
-function formatLastEditedTime(lastEditedTime?: string | null) {
-  if (!lastEditedTime) {
-    return '최근 수정일 정보 없음';
-  }
-
-  const date = new Date(lastEditedTime);
-
-  if (Number.isNaN(date.getTime())) {
-    return lastEditedTime;
-  }
-
-  const year = date.getFullYear();
-  const month = String(date.getMonth() + 1).padStart(2, '0');
-  const day = String(date.getDate()).padStart(2, '0');
-
-  return `${year}.${month}.${day}`;
 }
